requirements.txt
matplotlib>=3.8
numpy>=1.24
Pillow>=10.0

